+++ src/download.jl
@@ -17,9 +17,10 @@
         update::Bool = false,
         logfile::String = "downloads.log",
         loglevel::Symbol = :Debug
-   )
+    ) -> SortedDict{String,Any}
 
-Download satellite data from the Aeris/ICARE server.
+Download satellite data from the Aeris/ICARE server. The function returns a dictionary with
+the inventory of available online data for the given product.
 
 !!! note
     To use `sftp_download`, an [Aeris/ICARE account](https://www.icare.univ-lille.fr) is needed
@@ -47,7 +48,7 @@ or a year (if both day and month are omitted).
 - `convert::Bool`: Whether or not to convert the downloaded files to another file format (default: `true`).
 - `resync::Bool`: Whether to re-synchronize the local inventory with the remote server (default: `false`).
 - `update::Bool`: Whether to update the local files if newer versions are available
-  on the remote server (default: `false`).
+  on the remote server (default: `false`). Converted file sizes will be deleted for any updates.
 - `logfile::String`: The name of the log file (default: `"downloads.log"`; the name will be appended
   by the current date and time).
 - `loglevel::Symbol`: The log level for the download process (default: `:Debug`).
@@ -78,7 +79,7 @@ function sftp_download(
     update::Bool = false,
     logfile::String = "downloads.log",
     loglevel::Symbol = :Debug
-)::Nothing
+)::SortedDict{String,Any}
     ## Setup
     # Create product folder, if not existent
     product = isnothing(version) ? product : @sprintf("%s.v%.2f", product, version)
@@ -94,7 +95,9 @@ function sftp_download(
     open(logfile, "w") do logio
         logger = Logging.ConsoleLogger(logio, level, show_limited=false)
         Logging.with_logger(logger) do
-            @info "downloading '$product' data to '$(realpath(localroot))'"
+            range = daterange.start == daterange.stop ? daterange.start :
+                string(daterange.start, " – ", daterange.stop)
+            @info "downloading '$product' data to '$(realpath(localroot))' for $range"
         end
         #* Syncing local and remote database
         # Get connection to server, go to product folder on remote
@@ -104,7 +107,7 @@ function sftp_download(
         end
         icare = icare_connect(user, password, remoteroot, product, logger)
         # ℹ Make inventory available for catch block
-        inventory = OrderedDict{String,Any}()
+        inventory = SortedDict{String,Any}()
         # Get available server dates
         try
             product_database!(icare, inventory, localroot, product, daterange, convert, resync, logger)
@@ -116,9 +119,10 @@ function sftp_download(
             Logging.with_logger(logger) do
                 @error "failed to load local inventory" error
             end
+            data_gaps!(inventory)
             save_inventory(inventory, ts)
             @error "failed to load local inventory" _module=nothing _file=nothing _line=nothing
-            return
+            return inventory
         end
         # Log download session
         t0 = Dates.now()
@@ -137,13 +141,18 @@ function sftp_download(
         counter = Counter()
         # Match folder structure with server
         try sync!(icare, inventory, daterange, convert, update, resync, logger, logio, counter)
-        catch
+        catch error
+            Logging.with_logger(logger) do
+                @error "failed to sync with ICARE server" error
+            end
             @error "failed to sync with ICARE server" _module=nothing _file=nothing _line=nothing
         finally
             #* Log end of download session and save inventory
-            save_inventory(inventory, t0)
+            save_inventory(inventory, ts)
             log_counter(counter, logger, logio, t0)
             @info "download session closed"
+            # Return inventory for further investigation after download
+            return inventory
         end
     end #logging to file
 end #function ftp_download
@@ -281,7 +290,7 @@ end
 """
     sync!(
         icare::SFTP.Client,
-        inventory::OrderedDict,
+        inventory::SortedDict,
         daterange::@NamedTuple{start::Date, stop::Date},
         convert::Bool,
         update::Bool,
@@ -299,7 +308,7 @@ in the `logio` I/O stream.
 """
 function sync!(
     icare::SFTP.Client,
-    inventory::OrderedDict,
+    inventory::SortedDict,
     daterange::@NamedTuple{start::Date, stop::Date},
     convert::Bool,
     update::Bool,
@@ -309,8 +318,8 @@ function sync!(
     counter::Counter
 )::Nothing
     #* Define all files for download
-    dates = inventory["dates"].keys |> filter(d -> daterange.start ≤ d ≤ daterange.stop)
-    files = vcat([File.(Ref(icare), Ref(inventory), date, inventory["dates"][date].keys, convert)
+    dates = collect(Date, keys(inventory["dates"])) |> filter(d -> daterange.start ≤ d ≤ daterange.stop)
+    files = vcat([File.(Ref(icare), Ref(inventory), date, collect(String, keys(inventory["dates"][date])), convert)
         for date in dates]...)
 
     prog = pm.Progress(length(files), desc="downloading...")
@@ -329,12 +338,13 @@ function sync!(
             continue
         end
         t0 = Dates.now()
-        orig = isfile(file.location.download)
+        orig = isfile(file.location.download) && filesize(file.location.download) == inventory["dates"][file.date][file.name]["size"]
         #* Download file and optionally convert to another format
         try
             download(icare, inventory, file, update)
             convert!(inventory, file, convert, logger)
         catch error
+            rethrow(error)
             lock(thread) do
                 #* Log download errors
                 Logging.with_logger(logger) do
@@ -345,8 +355,9 @@ function sync!(
         #* Error handling/Re-download, if unsuccessful
         if !downloaded(inventory, file, update)
             # Check connection to ICARE server
-            icare = icare_connect(icare.username, icare.password, inventory["metadata"]["server"]["root"],
-                inventory["metadata"]["server"]["product"], logger)
+            @info "second download attempt for $(file.name)"
+            icare = icare_connect(icare.username, icare.password, inventory["metadata"]["remote"]["root"],
+                inventory["metadata"]["remote"]["product"], logger)
             # Check for correct server-side file stats
             update_stats!(icare, inventory, file, resync, logger)
             try
@@ -386,15 +397,15 @@ function sync!(
                     # Log successful downloads
                     counter.downloads += 1
                     Logging.with_logger(logger) do
-                        @debug("downloaded $(file.name) in $(Dates.canonicalize(t1 - t0)) @$t1",
-                            _module=nothing, _file=nothing, _line=nothing)
+                        fsize = inventory["dates"][file.date][file.name]["size"]
+                        msg = @sprintf("%s: downloaded %0.2f MB in %s with %0.2f MB/s @%s", file.name,
+                            fsize / 1e6, Dates.canonicalize(t1 - t0), fsize / (t1 - t0).value / 1e3, t1)
+                        @debug msg _module=nothing _file=nothing _line=nothing
                     end
                 end
             end
         end
-        lock(thread) do
-            flush(logio)
-        end
+        flush(logio)
         pm.next!(prog) # Update progress meter
     end # loop over files
     pm.finish!(prog)
@@ -404,7 +415,7 @@ end
 """
     download!(
         icare::SFTP.Client,
-        inventory::OrderedDict,
+        inventory::SortedDict,
         file::File,
         update
     )
@@ -414,11 +425,15 @@ Only download newer files on the server, if `update` is set to `true`.
 """
 function download(
     icare::SFTP.Client,
-    inventory::OrderedDict,
+    inventory::SortedDict,
     file::File,
     update
 )::Nothing
+    # Check, if file was already downloaded
     downloaded(inventory, file, update, true) && return
+    # Remove incomplete downloads
+    rm(file.location.download, force=true)
+    # Download file from server
     SFTP.download(icare, file.location.remote, file.dir.dst, force=true)
     return
 end
@@ -426,7 +441,7 @@ end
 
 """
     downloaded(
-        inventory::OrderedDict,
+        inventory::SortedDict,
         file::File,
         update::Bool,
         orig::Bool=false
@@ -438,20 +453,20 @@ is set, downloaded returns `false`, if newer versions of the `file` exist on the
 When `orig` is set to `true`, downloaded checks against the downloaded instead of the target file.
 """
 function downloaded(
-    inventory::OrderedDict,
+    inventory::SortedDict,
     file::File,
     update::Bool,
     orig::Bool=false
 )::Bool
     filestats = inventory["dates"][file.date][file.name]
     # Check, if file exists (orig: checks download of original file from server, not converted file)
-    file = orig ? file.location.download : file.location.target
-    isfile(file) || return false
+    dbfile = orig ? file.location.download : file.location.target
+    isfile(dbfile) || return false
     # Get file stats and type
-    localstats = stat(file)
-    size = if splitext(file)[2] == ".h5" && inventory["metadata"]["file"]["ext"] ≠ ".h5" # h5 converted files
+    localstats = stat(dbfile)
+    size = if splitext(dbfile)[2] == ".h5" && inventory["metadata"]["file"]["ext"] ≠ ".h5" # h5 converted files
         # ℹ Compare h5 size or return false for unknown h5 size in inventory
-        haskey(filestats, "converted") && (localstats.size == filestats["converted"])
+        haskey(filestats, "size"*file.newext) && (localstats.size == filestats["size"*file.newext])
     else # original file from server
         localstats.size == filestats["size"]
     end
@@ -466,7 +481,7 @@ end
 
 """
     convert!(
-        inventory::OrderedDict,
+        inventory::SortedDict,
         file::File,
         convert::Bool,
         logged::Logging.ConsoleLogger
@@ -476,7 +491,7 @@ Convert the `file` to a new file format as defined in the `inventory` unless `fi
 up-to-date or it was opted out to `convert` the file. Log events to `logger`.
 """
 function convert!(
-    inventory::OrderedDict,
+    inventory::SortedDict,
     file::File,
     convert::Bool,
     logger::Logging.ConsoleLogger
@@ -489,15 +504,15 @@ end
 
 
 """
-    converted!(inventory::OrderedDict, file::File, convert::Bool) -> Bool
+    converted!(inventory::SortedDict, file::File, convert::Bool) -> Bool
 
 Check, whether the size of the converted `file` is known in the `inventory` and matches the
 actual file size. Return also `true`, if it was opted out to `convert` the file.
 """
-function converted!(inventory::OrderedDict, file::File, convert::Bool)::Bool
-    if convert && haskey(inventory["dates"][file.date][file.name], "converted")
+function converted!(inventory::SortedDict, file::File, convert::Bool)::Bool
+    if convert && haskey(inventory["dates"][file.date][file.name], "size"*file.newext)
         # Compare file size with inventory
-        inventory["dates"][file.date][file.name]["converted"] == filesize(file.location.target)
+        inventory["dates"][file.date][file.name]["size"*file.newext] == filesize(file.location.target)
     else
         return !convert
     end
@@ -506,7 +521,7 @@ end
 
 """
     set_converted_size!(
-        inventory::OrderedDict,
+        inventory::SortedDict,
         file::File,
         convert::Bool,
         logger::Logging.ConsoleLogger
@@ -516,14 +531,14 @@ Set the size of the converted `file` in the `inventory` and mark the `inventory`
 Log events to `logger`.
 """
 function set_converted_size!(
-    inventory::OrderedDict,
+    inventory::SortedDict,
     file::File,
     convert::Bool,
     logger::Logging.ConsoleLogger
 )::Nothing
     # Initial checks
     convert || return
-    haskey(inventory["dates"][file.date][file.name], "converted") && return
+    haskey(inventory["dates"][file.date][file.name], "size"*file.newext) && return
     if !isfile(file.location.target)
         lock(thread) do
             # Log error, if converted file does not exist
@@ -536,7 +551,7 @@ function set_converted_size!(
     end
     # Save converted file size to inventory
     lock(thread) do
-        inventory["dates"][file.date][file.name]["converted"] = filesize(file.location.target)
+        inventory["dates"][file.date][file.name]["size"*file.newext] = filesize(file.location.target)
         inventory["metadata"]["database"]["updated"] = Dates.now()
     end
     return